project/requirements.txt
# This project uses python 2.7x
celery==4.0.2
Django==1.11
django-cors-headers==2.0.2
django-debug-toolbar==1.7
django-unslashed==0.3.0
flower==0.9.1
gunicorn==19.7.1
redis==2.10.5
psycopg2==2.7.1
requests==2.13.0
whitenoise==3.3.0
